# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Auto-Pay Cancel Assistant API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export interface SuccessResponse {
  success: boolean;
  message: string;
}

export interface RegisterRequest {
  email: string;
  /** @minLength 8 */
  password: string;
  name: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export type UserSubscriptionStatus =
  (typeof UserSubscriptionStatus)[keyof typeof UserSubscriptionStatus];

export const UserSubscriptionStatus = {
  trial: "trial",
  active: "active",
  cancelled: "cancelled",
  expired: "expired",
} as const;

export interface User {
  id: number;
  email: string;
  name: string;
  createdAt: string;
  trialEndsAt?: string;
  subscriptionStatus: UserSubscriptionStatus;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface BankAccount {
  id: number;
  bankName: string;
  accountType: string;
  lastFour: string;
  isActive: boolean;
  connectedAt: string;
}

export type ConnectBankRequestAccountType =
  (typeof ConnectBankRequestAccountType)[keyof typeof ConnectBankRequestAccountType];

export const ConnectBankRequestAccountType = {
  checking: "checking",
  savings: "savings",
  credit: "credit",
} as const;

export interface ConnectBankRequest {
  bankName: string;
  accountType: ConnectBankRequestAccountType;
  lastFour: string;
}

export type RecurringPaymentFrequency =
  (typeof RecurringPaymentFrequency)[keyof typeof RecurringPaymentFrequency];

export const RecurringPaymentFrequency = {
  weekly: "weekly",
  monthly: "monthly",
  quarterly: "quarterly",
  annually: "annually",
} as const;

export type RecurringPaymentStatus =
  (typeof RecurringPaymentStatus)[keyof typeof RecurringPaymentStatus];

export const RecurringPaymentStatus = {
  active: "active",
  cancelled: "cancelled",
  disputed: "disputed",
} as const;

export type RecurringPaymentCancellationDifficulty =
  (typeof RecurringPaymentCancellationDifficulty)[keyof typeof RecurringPaymentCancellationDifficulty];

export const RecurringPaymentCancellationDifficulty = {
  easy: "easy",
  medium: "medium",
  hard: "hard",
} as const;

export interface RecurringPayment {
  id: number;
  merchantName: string;
  amount: number;
  currency: string;
  frequency: RecurringPaymentFrequency;
  category: string;
  status: RecurringPaymentStatus;
  nextChargeDate: string;
  lastChargeDate?: string;
  accountId: number;
  cancellationDifficulty: RecurringPaymentCancellationDifficulty;
  logoUrl?: string;
}

export type UpdatePaymentStatusRequestStatus =
  (typeof UpdatePaymentStatusRequestStatus)[keyof typeof UpdatePaymentStatusRequestStatus];

export const UpdatePaymentStatusRequestStatus = {
  active: "active",
  cancelled: "cancelled",
  disputed: "disputed",
} as const;

export interface UpdatePaymentStatusRequest {
  status: UpdatePaymentStatusRequestStatus;
}

export type CancellationWorkflowDifficulty =
  (typeof CancellationWorkflowDifficulty)[keyof typeof CancellationWorkflowDifficulty];

export const CancellationWorkflowDifficulty = {
  easy: "easy",
  medium: "medium",
  hard: "hard",
} as const;

export type WorkflowStepActionType =
  (typeof WorkflowStepActionType)[keyof typeof WorkflowStepActionType];

export const WorkflowStepActionType = {
  navigate: "navigate",
  click: "click",
  fill: "fill",
  call: "call",
  email: "email",
  chat: "chat",
} as const;

export interface WorkflowStep {
  stepNumber: number;
  title: string;
  description: string;
  actionType: WorkflowStepActionType;
  actionUrl?: string;
}

export interface CancellationWorkflow {
  merchantName: string;
  difficulty: CancellationWorkflowDifficulty;
  steps: WorkflowStep[];
  cancellationUrl?: string;
  phoneNumber?: string;
  estimatedTime?: string;
  tips?: string[];
}

export interface GenerateEmailRequest {
  paymentId: number;
  userName: string;
  userEmail: string;
}

export interface GenerateDocumentRequest {
  paymentId: number;
  userName: string;
  bankName: string;
  accountLastFour: string;
}

export interface EmailTemplate {
  subject: string;
  body: string;
  recipientEmail?: string;
}

export interface GeneratedDocument {
  title: string;
  content: string;
  instructions: string[];
}

export type ActivityItemType =
  (typeof ActivityItemType)[keyof typeof ActivityItemType];

export const ActivityItemType = {
  cancelled: "cancelled",
  detected: "detected",
  disputed: "disputed",
  saved: "saved",
} as const;

export interface ActivityItem {
  id: number;
  type: ActivityItemType;
  merchantName: string;
  amount: number;
  date: string;
  description: string;
}

export interface DashboardSummary {
  totalMonthlySpend: number;
  activeSubscriptions: number;
  cancelledSubscriptions: number;
  potentialMonthlySavings: number;
  totalSaved: number;
  connectedAccounts: number;
  recentActivity: ActivityItem[];
}

export type GetRecurringPaymentsParams = {
  status?: GetRecurringPaymentsStatus;
  accountId?: number;
};

export type GetRecurringPaymentsStatus =
  (typeof GetRecurringPaymentsStatus)[keyof typeof GetRecurringPaymentsStatus];

export const GetRecurringPaymentsStatus = {
  active: "active",
  cancelled: "cancelled",
  disputed: "disputed",
} as const;