# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Auto-Pay Cancel Assistant API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /accounts:
    get:
      operationId: getBankAccounts
      tags: [accounts]
      summary: Get connected bank accounts
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of bank accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BankAccount"
    post:
      operationId: connectBankAccount
      tags: [accounts]
      summary: Connect a bank account (simulate Plaid)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConnectBankRequest"
      responses:
        "201":
          description: Bank account connected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BankAccount"
  /accounts/{accountId}:
    delete:
      operationId: disconnectBankAccount
      tags: [accounts]
      summary: Disconnect a bank account
      security:
        - bearerAuth: []
      parameters:
        - name: accountId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Account disconnected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /documents/email-template:
    post:
      operationId: generateEmailTemplate
      tags: [documents]
      summary: Generate cancellation email template
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateEmailRequest"
      responses:
        "200":
          description: Generated email template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailTemplate"
  /documents/ach-revocation:
    post:
      operationId: generateAchRevocation
      tags: [documents]
      summary: Generate ACH revocation letter
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateDocumentRequest"
      responses:
        "200":
          description: Generated ACH revocation letter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneratedDocument"
  /documents/stop-payment:
    post:
      operationId: generateStopPayment
      tags: [documents]
      summary: Generate stop payment request form
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateDocumentRequest"
      responses:
        "200":
          description: Generated stop payment form
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneratedDocument"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message
    RegisterRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        name:
          type: string
      required:
        - email
        - password
        - name
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - token
        - user
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        createdAt:
          type: string
          format: date-time
        trialEndsAt:
          type: string
          format: date-time
        subscriptionStatus:
          type: string
          enum: [trial, active, cancelled, expired]
      required:
        - id
        - email
        - name
        - createdAt
        - subscriptionStatus
    BankAccount:
      type: object
      properties:
        id:
          type: integer
        bankName:
          type: string
        accountType:
          type: string
        lastFour:
          type: string
        isActive:
          type: boolean
        connectedAt:
          type: string
          format: date-time
      required:
        - id
        - bankName
        - accountType
        - lastFour
        - isActive
        - connectedAt
    ConnectBankRequest:
      type: object
      properties:
        bankName:
          type: string
        accountType:
          type: string
          enum: [checking, savings, credit]
        lastFour:
          type: string
      required:
        - bankName
        - accountType
        - lastFour
...
    GenerateEmailRequest:
      type: object
      properties:
        paymentId:
          type: integer
        userName:
          type: string
        userEmail:
          type: string
      required:
        - paymentId
        - userName
        - userEmail
    GenerateDocumentRequest:
      type: object
      properties:
        paymentId:
          type: integer
        userName:
          type: string
        bankName:
          type: string
        accountLastFour:
          type: string
      required:
        - paymentId
        - userName
        - bankName
        - accountLastFour
    EmailTemplate:
      type: object
      properties:
        subject:
          type: string
        body:
          type: string
        recipientEmail:
          type: string
      required:
        - subject
        - body
    GeneratedDocument:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        instructions:
          type: array
          items:
            type: string
      required:
        - title
        - content
        - instructions
...