# lib/api-zod/src/generated/types/generateDocumentRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Auto-Pay Cancel Assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface GenerateDocumentRequest {
  paymentId: number;
  userName: string;
  bankName: string;
  accountLastFour: string;
}