# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Auto-Pay Cancel Assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { UserSubscriptionStatus } from "./userSubscriptionStatus";

export interface User {
  id: number;
  email: string;
  name: string;
  createdAt: Date;
  trialEndsAt?: Date;
  subscriptionStatus: UserSubscriptionStatus;
}